move wavetable offset commands, proper oscilloscope, fixed noise pitch, 1-bit noise and command for switching noise modes, more fm volume functions, proper oscilloscope image (it displayed inverted wave), some bug fixes, updated wavegen

+++ src/wavegen.h
@@ -29,7 +29,7 @@ typedef struct
 	int mult, shift;
 	int exp;
 	
-	Uint8 vol; //wasn't there
+	Uint16 vol; //wasn't there
 	
 	float exp_c;
 	Uint32 flags;
